feat: restructuring of folders
